Move class to test folder b/c it's only used in tests
