+++ src/test/java/org/cru/godtools/utils/ImageReader.java
@@ -1,4 +1,4 @@
-package org.cru.godtools.api.utilities;
+package org.cru.godtools.utils;
 
 import com.google.common.base.Throwables;
